requirements.txt
numpy
streamlit
pillow
matplotlib
scikit-learn
tqdm
albumentations
torch
torchvision
torchaudio
opencv-python

